Override default location of public dir in prod
Strapi v4 lost the ability to override the public directory. v4.1.6
re-added that functionality but in a different way than how it worked in
Strapi pre v4. This PR[1] moves the configuration of the location of the
public directory to config/server.js. We want to override the default
only in production so that the public folder contents can be saved on a
persistent disk.

[1]: https://github.com/strapi/strapi/pull/12534 Although this PR was
merged on Mar 25, the change wasn't included in a release until 4.1.6,
which was released on Mar 30.